Split Request mapping
