+++ browser/request_mapping.go
@@ -1,0 +1,36 @@
+package browser
+
+import (
+	"github.com/dop251/goja"
+
+	"github.com/grafana/xk6-browser/common"
+)
+
+// mapRequest to the JS module.
+func mapRequest(vu moduleVU, r *common.Request) mapping {
+	rt := vu.Runtime()
+	maps := mapping{
+		"allHeaders": r.AllHeaders,
+		"frame": func() *goja.Object {
+			mf := mapFrame(vu, r.Frame())
+			return rt.ToValue(mf).ToObject(rt)
+		},
+		"headerValue":         r.HeaderValue,
+		"headers":             r.Headers,
+		"headersArray":        r.HeadersArray,
+		"isNavigationRequest": r.IsNavigationRequest,
+		"method":              r.Method,
+		"postData":            r.PostData,
+		"postDataBuffer":      r.PostDataBuffer,
+		"resourceType":        r.ResourceType,
+		"response": func() *goja.Object {
+			mr := mapResponse(vu, r.Response())
+			return rt.ToValue(mr).ToObject(rt)
+		},
+		"size":   r.Size,
+		"timing": r.Timing,
+		"url":    r.URL,
+	}
+
+	return maps
+}
